refactor state and line
The state.c and line.c files contained code for parsing and formatting
assembler, but all of this code is only needed by specasm.
Saexport and saimport only need the formatting and parsing code
respectively.  By splitting things up in this way we reduce
the size of saimport and saexport and speed up compile times.
This change is also necessary as the addition of the expression
syntax meant that saimport could no longer be built, without
excluding the code it didn't need.
