+++ src/line_common.c
@@ -1,0 +1,102 @@
+/*
+ * Copyright contributors to Specasm
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+
+#include "line_common.h"
+
+char byte_regs[8] = {'b', 'c', 'd', 'e', 'h', 'l', ' ', 'a'};
+
+/*
+ * Must be in the same order as the opcode_table in line_parse.c.
+ */
+
+const specasm_mnemomic_t mnemomics_table[] = {
+	{ "adc", SPECASM_LINE_TYPE_ADC, },
+	{ "add", SPECASM_LINE_TYPE_ADD, },
+	{ "align", SPECASM_LINE_TYPE_ALIGN, },
+	{ "and", SPECASM_LINE_TYPE_AND, },
+	{ "bit", SPECASM_LINE_TYPE_BIT, },
+	{ "call", SPECASM_LINE_TYPE_CALL, },
+	{ "ccf", SPECASM_LINE_TYPE_CCF, },
+	{ "cp", SPECASM_LINE_TYPE_CP, },
+	{ "cpd", SPECASM_LINE_TYPE_CPD, },
+	{ "cpdr", SPECASM_LINE_TYPE_CPDR },
+	{ "cpi", SPECASM_LINE_TYPE_CPI, },
+	{ "cpir", SPECASM_LINE_TYPE_CPIR, },
+	{ "cpl", SPECASM_LINE_TYPE_CPL, },
+	{ "daa", SPECASM_LINE_TYPE_DAA, },
+	{ "db", SPECASM_LINE_TYPE_DB, },
+	{ "dec", SPECASM_LINE_TYPE_DEC, },
+	{ "di", SPECASM_LINE_TYPE_DI, },
+	{ "djnz", SPECASM_LINE_TYPE_DJNZ, },
+	{ "ds", SPECASM_LINE_TYPE_DS, },
+	{ "dw", SPECASM_LINE_TYPE_DW, },
+	{ "ei", SPECASM_LINE_TYPE_EI, },
+	{ "ex", SPECASM_LINE_TYPE_EX, },
+	{ "exx", SPECASM_LINE_TYPE_EXX, },
+	{ "halt", SPECASM_LINE_TYPE_HALT, },
+	{ "im", SPECASM_LINE_TYPE_IM, },
+	{ "in", SPECASM_LINE_TYPE_IN },
+	{ "inc", SPECASM_LINE_TYPE_INC, },
+	{ "ind", SPECASM_LINE_TYPE_IND, },
+	{ "indr", SPECASM_LINE_TYPE_INDR, },
+	{ "ini", SPECASM_LINE_TYPE_INI, },
+	{ "inir", SPECASM_LINE_TYPE_INIR, },
+	{ "jp", SPECASM_LINE_TYPE_JP, },
+	{ "jr", SPECASM_LINE_TYPE_JR, },
+	{ "ld", SPECASM_LINE_TYPE_LD, },
+	{ "ldd", SPECASM_LINE_TYPE_LDD, },
+	{ "lddr", SPECASM_LINE_TYPE_LDDR, },
+	{ "ldi", SPECASM_LINE_TYPE_LDI, },
+	{ "ldir", SPECASM_LINE_TYPE_LDIR, },
+	{ "map", SPECASM_LINE_TYPE_MAP, },
+	{ "neg", SPECASM_LINE_TYPE_NEG, },
+	{ "nop", SPECASM_LINE_TYPE_NOP, },
+	{ "or", SPECASM_LINE_TYPE_OR, },
+	{ "org", SPECASM_LINE_TYPE_ORG, },
+	{ "otdr", SPECASM_LINE_TYPE_OTDR, },
+	{ "otir", SPECASM_LINE_TYPE_OTIR, },
+	{ "out", SPECASM_LINE_TYPE_OUT, },
+	{ "outd", SPECASM_LINE_TYPE_OUTD, },
+	{ "outi", SPECASM_LINE_TYPE_OUTI, },
+	{ "pop", SPECASM_LINE_TYPE_POP, },
+	{ "push", SPECASM_LINE_TYPE_PUSH, },
+	{ "res", SPECASM_LINE_TYPE_RES, },
+	{ "ret", SPECASM_LINE_TYPE_RET, },
+	{ "reti", SPECASM_LINE_TYPE_RETI, },
+	{ "retn", SPECASM_LINE_TYPE_RETN, },
+	{ "rl", SPECASM_LINE_TYPE_RL, },
+	{ "rla", SPECASM_LINE_TYPE_RLA, },
+	{ "rlc", SPECASM_LINE_TYPE_RLC, },
+	{ "rlca", SPECASM_LINE_TYPE_RLCA, },
+	{ "rld", SPECASM_LINE_TYPE_RLD, },
+	{ "rr", SPECASM_LINE_TYPE_RR, },
+	{ "rra", SPECASM_LINE_TYPE_RRA, },
+	{ "rrc", SPECASM_LINE_TYPE_RRC, },
+	{ "rrca", SPECASM_LINE_TYPE_RRCA, },
+	{ "rrd", SPECASM_LINE_TYPE_RRD, },
+	{ "rst", SPECASM_LINE_TYPE_RST, },
+	{ "sbc", SPECASM_LINE_TYPE_SBC, },
+	{ "scf", SPECASM_LINE_TYPE_SCF, },
+	{ "set", SPECASM_LINE_TYPE_SET, },
+	{ "sla", SPECASM_LINE_TYPE_SLA, },
+	{ "sra", SPECASM_LINE_TYPE_SRA, },
+	{ "srl", SPECASM_LINE_TYPE_SRL, },
+	{ "sub", SPECASM_LINE_TYPE_SUB, },
+	{ "xor", SPECASM_LINE_TYPE_XOR, },
+};
+
+const uint8_t mnemomics_table_size =
+	sizeof(mnemomics_table) / sizeof(const specasm_mnemomic_t);
