+++ src/state_parse.c
@@ -426,112 +426,3 @@ void specasm_parse_line_e(unsigned int l, const char *str)
 	prv_parse_short_comment_e(str, i + 1, line);
 }
 
-static char *prv_format_string_e(char *buf, uint8_t lng, unsigned int l,
-				 uint8_t ch)
-{
-	const char *ptr;
-
-	if (lng)
-		ptr = specasm_state_get_long_e(l);
-	else
-		ptr = specasm_state_get_short_e(l);
-	if (err_type != SPECASM_ERROR_OK)
-		return NULL;
-	*buf++ = ch;
-	while (*ptr)
-		*buf++ = *ptr++;
-
-	return buf;
-}
-
-char* prv_format_equ_e(char *buf, const specasm_line_t *line)
-{
-	const uint8_t *op_code = &line->data.op_code[0];
-	buf = prv_format_string_e(buf, op_code[0] == SPECASM_LINE_TYPE_LL,
-				    op_code[1], '.');
-	if (err_type != SPECASM_ERROR_OK)
-		return NULL;
-
-	buf[0] = ' ';
-	buf[1] = 'E';
-	buf[2] = 'Q';
-	buf[3] = 'U';
-
-	return prv_format_string_e(buf + 4, op_code[2] == SPECASM_LINE_TYPE_LL,
-				    op_code[3], ' ');
-}
-
-void specasm_format_line_e(char *buf, unsigned int l)
-{
-	const char *ptr;
-	const char *end_ptr;
-	char *comment_start;
-	char *start;
-	uint8_t i;
-	const char str_ids[] = {'\'', '"', '#', '@', '-', '+'};
-	const specasm_line_t *line = &state.lines.lines[l];
-
-	end_ptr = buf + SPECASM_LINE_MAX_LEN;
-
-	if (line->type == SPECASM_LINE_TYPE_EMPTY)
-		goto clear;
-
-	start = buf;
-
-	if (line->type == SPECASM_LINE_TYPE_EQU) {
-		buf = prv_format_equ_e(buf, line);
-		goto clear;
-	}
-
-	if ((line->type == SPECASM_LINE_TYPE_LL) ||
-	    (line->type == SPECASM_LINE_TYPE_SL)) {
-		buf =
-		    prv_format_string_e(buf, line->type == SPECASM_LINE_TYPE_LL,
-					line->data.label, '.');
-		goto clear;
-	}
-
-	if ((line->type == SPECASM_LINE_TYPE_LC) ||
-	    (line->type == SPECASM_LINE_TYPE_SC)) {
-		buf =
-		    prv_format_string_e(buf, line->type == SPECASM_LINE_TYPE_LC,
-					line->comment, ';');
-		goto clear;
-	}
-	if ((line->type >= SPECASM_LINE_TYPE_STR_SIN_SHORT) &&
-	    (line->type <= SPECASM_LINE_TYPE_INC_SYS_LONG)) {
-		i = (line->type - SPECASM_LINE_TYPE_STR_SIN_SHORT) >> 1;
-		i = str_ids[i];
-		buf = prv_format_string_e(buf, line->type & 1, line->data.label,
-					  i);
-		if ((line->type < SPECASM_LINE_TYPE_INC_SHORT) && buf < end_ptr)
-			*buf++ = i;
-	} else {
-		if (!((line->type >= SPECASM_LINE_TYPE_DB &&
-		       line->type <= SPECASM_LINE_TYPE_DW_SUB) ||
-		      line->type == SPECASM_LINE_TYPE_DS))
-			for (i = 0; i < SPECASM_MAX_INDENT; i++)
-				*buf++ = ' ';
-
-		buf += specasm_dump_opcode_e(line, buf);
-		if (err_type != SPECASM_ERROR_OK)
-			goto clear;
-	}
-
-	if (line->comment != SPECASM_NULL) {
-		comment_start = start + SPECASM_LINE_MAX_OPCODE + 1;
-		while (buf < comment_start)
-			*buf++ = ' ';
-		*buf++ = ';';
-		ptr = specasm_state_get_short_e(line->comment);
-		if (err_type != SPECASM_ERROR_OK)
-			goto clear;
-		while (*ptr)
-			*buf++ = *ptr++;
-	}
-
-clear:
-	while (buf < end_ptr)
-		*buf++ = ' ';
-	*buf = 0;
-}
